move TopKConsumer functions into that class
